Done: 1.ng-tree-copy-paste done. 2.discard delete-object in ng-tree. 3. merge upload & create page in editor page. 4.change JSOGRef/JSOGGenerator ID into Time, the type int->long, then cast to String. 5.childGen() in ng-tree component add array of value. 6.add ng-tree expandAll() initial and update Data to sessionStorage. 7.generate-form just have save/clearform function, move clear-sessionStorage to ng-tree. 8.move ng-tree-component to create.html, and this is named Editor in app-router and view. Need: 1.img upload/store/view. 2.validation. 3.Apache CLI.

+++ src/main/java/jetty/demo/JSOGRef.java
@@ -41,9 +41,11 @@ public class JSOGRef
 		type = inputType;
 	}
 
-	/** */
-	public JSOGRef(int val, String inputType) {
-		this(Integer.toString(val), inputType);
+	/** 
+	 * change int argument to long for date.getTime()
+	 * */
+	public JSOGRef(long val, String inputType) {
+		this(Long.toString(val), inputType);
 	}
 
 	@Override
